+++ resources/views/components/layout/navbar.blade.php
@@ -73,11 +73,16 @@
             <nav aria-label="Global" class="flex space-x-4">
                 {{-- <a href="{{ route('schedule') }}"
                     class="text-sm font-medium text-gray-900 p-2 hover:bg-gray-100 rounded-lg">Courses</a> --}}
-                {{-- <a href="#" class="text-sm font-medium text-gray-900 p-2 hover:bg-gray-100 rounded-lg">
-                    Events
-                </a> --}}
-                {{-- <a href="#" class="text-sm font-medium text-gray-900 p-2 hover:bg-gray-100 rounded-lg">Blog</a>
-                --}}
+
+                @guest
+                <a href="{{ route('register') }}"
+                    class="text-sm font-medium text-gray-500 p-2 hover:bg-gray-100 rounded-lg">Register</a>
+                <a href="{{ route('login') }}"
+                    class="inline-flex items-center px-3 py-2 border border-transparent text-sm leading-4 font-medium rounded-md text-indigo-700 bg-indigo-100 hover:bg-indigo-200 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
+                    Login
+                </a>
+                @endguest
+
             </nav>
             <div class="flex items-center space-x-8">
                 {{-- <span class="inline-flex">