+++ resources/views/course/dashboard/index.blade.php
@@ -19,7 +19,7 @@
         </div>
     </x-slot>
 
-    <div class="max-h-screen">
+    <div class="h-screen overflow-y-scroll">
         <div class="max-w-3xl mx-auto sm:px-6 lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-12 lg:gap-8">
             <div class="hidden lg:block lg:col-span-3 xl:col-span-2 py-4 sm:py-6 md:py-8 lg:py-10">
                 @include('course.dashboard.left')