+++ resources/views/welcome.blade.php
@@ -38,31 +38,6 @@
                     <p class="mt-3 max-w-md mx-auto text-base text-gray-500 sm:text-lg md:mt-5 md:text-xl md:max-w-3xl">
                         All dancing events in Croatia and around in one place.
                     </p>
-                    <div class="mt-5 max-w-md mx-auto sm:flex sm:justify-center md:mt-8">
-                        <div class="rounded-md">
-                            {{-- <a href="/auth/redirect"
-                                class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 md:py-4 md:text-lg md:px-10">
-                                <span class="sr-only">Sign in with Facebook</span>
-                                <svg class="w-5 h-5" fill="currentColor" viewBox="0 0 20 20" aria-hidden="true">
-                                    <path fill-rule="evenodd"
-                                        d="M20 10c0-5.523-4.477-10-10-10S0 4.477 0 10c0 4.991 3.657 9.128 8.438 9.878v-6.987h-2.54V10h2.54V7.797c0-2.506 1.492-3.89 3.777-3.89 1.094 0 2.238.195 2.238.195v2.46h-1.26c-1.243 0-1.63.771-1.63 1.562V10h2.773l-.443 2.89h-2.33v6.988C16.343 19.128 20 14.991 20 10z"
-                                        clip-rule="evenodd" />
-                                </svg>
-                                <span class="ml-2 text-sm">Login with Facebook</span>
-                            </a> --}}
-                            {{-- <a href=""
-                                class="w-full inline-flex justify-center py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-indigo-100 bg-indigo-600 hover:bg-indigo-700">
-
-                            </a> --}}
-                        </div>
-                        <div class="mt-3 rounded-md shadow sm:mt-0 sm:ml-3">
-                            {{-- <a href="#main"
-                                class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-indigo-600 bg-white hover:bg-gray-50 md:py-4 md:text-lg md:px-10">
-                                @include('icons.events')
-                                <span class="ml-2 text-sm">Schedule</span>
-                            </a> --}}
-                        </div>
-                    </div>
                 </div>
             </main>
         </div>
@@ -70,7 +45,7 @@
     @endguest
 
     <!-- This example requires Tailwind CSS v2.0+ -->
-    <div class="bg-gray-100">
+    <div class="bg-indigo-50">
         <div class="max-w-7xl mx-auto text-center py-12 px-4 sm:px-6 lg:py-16 lg:px-8">
             <h2 class="text-3xl font-extrabold tracking-tight text-gray-900 sm:text-4xl">
                 <span class="block">Ready to start dancing?</span>